% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_polygon.R
\name{pb_weightedPolyJoin}
\alias{pb_weightedPolyJoin}
\title{pb_weightedPolyJoin}
\usage{
pb_weightedPolyJoin(
  displaced.sf,
  polygons.sf,
  displaced.id = "DHSID",
  metrics = c("sri_score", "sri_basicamenities", "sri_basicequip", "sri_diagcapacity",
    "sri_infprev", "sri_med"),
  densityBuffer = NULL,
  adminBound = NULL,
  adminID = "ID_2",
  n.cores = 1
)
}
\arguments{
\item{displaced.sf}{Displaced coordinates in an sf object type.}

\item{polygons.sf}{Polygons to join the displaced points to. For now, the function assumes there is a column with unique ID for each polygon called "SPAID"}

\item{metrics}{Character vector of column names containing the numeric metrics you wish to calculate.}

\item{densityBuffer}{A density buffer list object created with the pb_integratedDens function.}

\item{adminBound}{(Optional) the administrative boundary that circumscribes the displacement.}

\item{adminID}{The unique ID for the adminBound features, defaults to "ID_2".}

\item{n.cores}{This allows for parallelization using the futures package. This can be slightly unstable, but normally functions well and dramatically speeds compute time when it does. n.cores specifies the number of cores to use. The default is 1, and does not parallelize.}
}
\value{
Returns a dataframe of length nrow(displaced.sf), with a DHSID column containing the unique ID of displaced.sf and SPAID containing the unique ID of the most likely polygon that a given point falls in.
}
\description{
For each given displaced community, estimates the raster value based on the probability weighted mean of possible true locations
}
\examples{

# coming soon!

}
\author{
J.W. Rozelle
}
