% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_rasterize.R
\name{st_rasterAsPoint}
\alias{st_rasterAsPoint}
\title{Convert probability raster to sf points with weight column "layer"}
\usage{
st_rasterAsPoint(probRaster, weightsCol = "layer")
}
\arguments{
\item{probRaster}{raster::RasterLayer}

\item{weightsCol}{name for weights column (default "layer")}
}
\value{
sf POINT with weights column
}
\description{
Internal helper used by join_* functions.
}
\keyword{internal}
