% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimProbBuff.R
\name{trimProbBuff}
\alias{trimProbBuff}
\title{trimProbBuff}
\usage{
trimProbBuff(probRaster = NULL, adminBound = NULL)
}
\arguments{
\item{probRaster}{Probability raster, object of class raster}

\item{adminBound}{The administrative / polygon boundary of class SF that the raster should be masked by}
}
\value{
Returns a trimmed object of class raster, with values summed to 1.
}
\description{
Trims and reweights probability buffer (from a raster object) by a polygon that it is limited to
}
\examples{

# coming soon!


}
\author{
J.W. Rozelle
}
