% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_count.R
\name{pb_countFeatures}
\alias{pb_countFeatures}
\title{pb_countFeatures}
\usage{
pb_countFeatures(
  displaced.sf,
  features2count.sf,
  radiusLength = 10000,
  displaced.id = "DHSID",
  densityBuffer = NULL,
  adminBound = NULL,
  adminID = NULL,
  n.cores = 1
)
}
\arguments{
\item{displaced.sf}{Displaced coordinates in an sf object type.}

\item{features2count.sf}{sf point feature that is counted.}

\item{radiusLength}{a numeric vector to determine the distance from a displaced community within which the number of features2count.sf will be counted.}

\item{densityBuffer}{A density buffer list object created with the pb_integratedDens function.}

\item{adminBound}{(Optional) the administrative boundary that circumscribes the displacement.}

\item{adminID}{The unique ID for the adminBound features, defaults to "ID_2".}

\item{n.cores}{This allows for parallelization using the futures package. This can be slightly unstable, but normally functions well and dramatically speeds compute time when it does. n.cores specifies the number of cores to use. The default is 1, and does not parallelize.}
}
\value{
a vector of the most likely count of point features within a defined radius of the displaced coordinates.
}
\description{
For each given displaced community, estimates the raster value based on the probability weighted mean of possible true locations
}
\examples{

# coming soon!




}
\author{
J.W. Rozelle
}
