% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_polygon.R
\name{pb_polyJoin}
\alias{pb_polyJoin}
\title{pb_polyJoin}
\usage{
pb_polyJoin(
  displaced.sf,
  polygons.sf,
  displaced.id = "DHSID",
  densityBuffer = NULL,
  adminBound = haiti_adm2,
  adminID = "ID_2",
  n.cores = 1
)
}
\arguments{
\item{displaced.sf}{Displaced coordinates in an sf object type.}

\item{polygons.sf}{Polygons to join the displaced points to. For now, the function assumes there is a column with unique ID for each polygon called "SPAID"}

\item{displaced.id}{The name of the column that contains a unique ID for the displaced communities.}

\item{densityBuffer}{A density buffer list object created with the pb_integratedDens or pb_Density functions.}

\item{adminBound}{(Optional) the administrative boundary that circumscribes the displacement}

\item{adminID}{The unique ID for the adminBound featurs, defaults to "ID_2".}

\item{n.cores}{This allows for parallelization using the futures package. This can be slightly unstable, but normally functions well and dramatically speeds compute time when it does. n.cores specifies the number of cores to use. The default is 1, and does not parallelize.}
}
\value{
Returns a dataframe of length nrow(displaced.sf), with a DHSID column containing the unique ID of displaced.sf and SPAID containing the unique ID of the most likely polygon that a given point falls in.
}
\description{
Estimates th
}
\examples{

# coming soon!


}
\author{
J.W. Rozelle
}
