% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_count.R
\name{pb_radiusLimitJoin}
\alias{pb_radiusLimitJoin}
\title{pb_radiusLimitJoin}
\usage{
pb_radiusLimitJoin(
  displaced.sf,
  features2count.sf,
  limitDist = 5000,
  displaced.id = "DHSID",
  features2count.id = "SPAID",
  metrics = c("sri_score", "sri_basicamenities", "sri_basicequip", "sri_diagcapacity",
    "sri_infprev", "sri_med"),
  densityBuffer = NULL,
  adminBound = NULL,
  adminID = NULL,
  weightedMedian = TRUE,
  probCuts = c(0.1, 0.9),
  n.cores = 1
)
}
\arguments{
\item{displaced.sf}{Displaced coordinates in an sf object type.}

\item{features2count.sf}{sf point feature that is counted.}

\item{limitDist}{a numeric vector to determine the distance from a displaced community within which the number of features2count.sf will be counted.}

\item{displaced.id}{a character vector of the column name specifying the unique ID of the displaced communities}

\item{features2count.id}{a charactor vector of the column name specifying the unique ID of the point features being connected}

\item{metrics}{Character vector of column names containing the numeric metrics you wish to calculate.}

\item{densityBuffer}{A density buffer list object created with the pb_integratedDens function.}

\item{adminBound}{(Optional) the administrative boundary that circumscribes the displacement.}

\item{adminID}{The unique ID for the adminBound features.}

\item{weightedMedian}{Boolean value, if true, the weighted median will be calculated. If false, the weighted mean will be calculated. Defaults to true.}

\item{probCuts}{A numeric vector containing values  between 0 and 1, these are percentiles to report the credible interval of distances}

\item{n.cores}{This allows for parallelization using the futures package. This can be slightly unstable, but normally functions well and dramatically speeds compute time when it does. n.cores specifies the number of cores to use. The default is 1, and does not parallelize.}
}
\value{
a vector of the most likely count of point features within a defined radius of the displaced coordinates.
}
\description{
For each given displaced community, estimates the raster value based on the probability weighted mean of possible true locations
}
\examples{

# coming soon!




}
\author{
J.W. Rozelle
}
